move filtering to proposal list
